+++ examples/workflow-router/CHANGELOG.md
@@ -1,5 +1,84 @@
 # Changelog
 
+## [1.17.7](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.6...example-workflow-router-v1.17.7) (2025-07-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.22.6
+
+## [1.17.6](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.5...example-workflow-router-v1.17.6) (2025-07-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.20.3
+    * @aigne/cli bumped to 1.22.5
+    * @aigne/core bumped to 1.33.0
+    * @aigne/openai bumped to 0.9.0
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.5
+
+## [1.17.5](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.4...example-workflow-router-v1.17.5) (2025-07-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/cli bumped to 1.22.4
+
+## [1.17.4](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.3...example-workflow-router-v1.17.4) (2025-07-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.20.2
+    * @aigne/cli bumped to 1.22.3
+    * @aigne/core bumped to 1.32.2
+    * @aigne/openai bumped to 0.8.2
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.4
+
+## [1.17.3](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.2...example-workflow-router-v1.17.3) (2025-07-09)
+
+
+### Bug Fixes
+
+* **model:** ensure last message is not system role for gemini ([#231](https://github.com/AIGNE-io/aigne-framework/issues/231)) ([1b72e1e](https://github.com/AIGNE-io/aigne-framework/commit/1b72e1e6be98060aa32e68585142b2eea401d109))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.20.1
+    * @aigne/cli bumped to 1.22.2
+    * @aigne/core bumped to 1.32.1
+    * @aigne/openai bumped to 0.8.1
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.3
+
+## [1.17.2](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.1...example-workflow-router-v1.17.2) (2025-07-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.20.0
+    * @aigne/cli bumped to 1.22.1
+    * @aigne/core bumped to 1.32.0
+    * @aigne/openai bumped to 0.8.0
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.2
+
 ## [1.17.1](https://github.com/AIGNE-io/aigne-framework/compare/example-workflow-router-v1.17.0...example-workflow-router-v1.17.1) (2025-07-04)