+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,46 @@
 # Changelog
 
+## [0.5.5](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.4...test-utils-v0.5.5) (2025-07-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.33.0
+
+## [0.5.4](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.3...test-utils-v0.5.4) (2025-07-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.32.2
+
+## [0.5.3](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.2...test-utils-v0.5.3) (2025-07-09)
+
+
+### Bug Fixes
+
+* **model:** ensure last message is not system role for gemini ([#231](https://github.com/AIGNE-io/aigne-framework/issues/231)) ([1b72e1e](https://github.com/AIGNE-io/aigne-framework/commit/1b72e1e6be98060aa32e68585142b2eea401d109))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.32.1
+
+## [0.5.2](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.1...test-utils-v0.5.2) (2025-07-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.32.0
+
 ## [0.5.1](https://github.com/AIGNE-io/aigne-framework/compare/test-utils-v0.5.0...test-utils-v0.5.1) (2025-07-04)